casestudy/requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
pypdf
pinecone
python-dotenv
streamlit